Initial code cleanup - Moved navigation things to navigation package - Cleaned up login stuff
