+++ app/src/main/java/com/example/stablestable/MainActivity.kt
@@ -9,7 +9,7 @@ import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
-import com.example.stablestable.ui.SetupNavGraph
+import com.example.stablestable.navigation.SetupNavGraph
 import com.example.stablestable.ui.theme.StableStableTheme
 
 class MainActivity : ComponentActivity() {
